--- src/main/java/utils/BlockStatus.java
@@ -1,7 +1,0 @@
-package utils;
-
-public enum BlockStatus {
-    OPERATIONAL,
-    BROKEN,
-    IN_REPAIR
-}
