--- src/main/java/utils/Suggestion.java
@@ -1,14 +1,0 @@
-package utils;
-
-import java.util.List;
-
-public class Suggestion {
-    public final int blockId;
-    public final int speed;
-    public final int[] authority;
-    public Suggestion(int b, int s, int[] a) {
-        blockId = b;
-        speed = s;
-        authority = a;
-    }
-}
